refactor: split productRouter into separate files, add 2 controllers, update routes
