+++ server/src/controllers/productController.ts
@@ -1,189 +1,7 @@
-/*
-
-model Product {
-  id          String             @id @default(cuid()) // "prod001"
-  name        String // "Смартфон Samsung Galaxy S24"
-  description String? // "Флагманский смартфон с топовыми характеристиками."
-  categoryId  String // "cat001"
-  category    Category           @relation(fields: [categoryId], references: [id])
-  images      String[] // ["https://example.com/samsung_s24.jpg"]
-  variants    ProductVariant[]
-  attributes  ProductAttribute[]
-  discountId  String? // "disc001"
-  discount    Discount?          @relation(fields: [discountId], references: [id])
-}
-
-model ProductAttribute {
-  id        String      @id @default(cuid()) // "attr001"
-  productId String // "prod001"
-  product   Product     @relation(fields: [productId], references: [id])
-  typeId    String // "varType002"
-  type      VariantType @relation(fields: [typeId], references: [id])
-  value     String // "Snapdragon 8 Gen 3"
-}
-
-model ProductVariant {
-  id         String                    @id @default(cuid()) // "var001"
-  productId  String // "prod001"
-  product    Product                   @relation(fields: [productId], references: [id])
-  sku        String                    @unique // "SAMSUNG-S24-256GB-BLACK"
-  price      Decimal // 89999.00
-  inStock    Int                       @default(10) // 10 штук на складе
-  attributes ProductVariantAttribute[]
-  discountId String? // null (нет скидки)
-  discount   Discount?                 @relation(fields: [discountId], references: [id])
-}
-
-model ProductVariantAttribute {
-  id        String         @id @default(cuid()) // "varAttr001"
-  variantId String // "var001"
-  variant   ProductVariant @relation(fields: [variantId], references: [id])
-  typeId    String // "varType001"
-  type      VariantType    @relation(fields: [typeId], references: [id])
-  value     String // "256 ГБ"
-}
-
-------------------- READY ---------------------------
-model VariantType {
-  id                      String                    @id @default(cuid()) // "varType003"
-  categoryId              String // "cat001"
-  category                Category                  @relation(fields: [categoryId], references: [id])
-  name                    String // "Производитель"
-  ProductAttribute        ProductAttribute[]
-  ProductVariantAttribute ProductVariantAttribute[]
-}
-*/
-
 import { PrismaClient } from '@prisma/client';
-import { Request, Response } from 'express';
+import { RouterHandler, ErrorHandler } from '../utils/controllerUtils';
 
 const prisma = new PrismaClient();
-
-type RouterHandler = (req: Request, res: Response) => Promise<void>;
-
-const ErrorHandler = (error: any, res: Response) => {
-    const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-    res.status(500).json({
-        error: 'Internal Server Error',
-        details: errorMessage,
-    });
-};
-
-export const getAllVariantTypes: RouterHandler = async (req, res) => {
-    try {
-        const variantTypes = await prisma.variantType.findMany();
-
-        res.status(200).json(variantTypes);
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const getVariantTypeById: RouterHandler = async (req, res) => {
-    try {
-        const { id } = req.params;
-        const variantType = await prisma.variantType.findUnique({
-            where: { id },
-        });
-
-        if (!variantType) {
-            res.status(404).json({ message: 'VariantType not found' });
-            return;
-        }
-
-        res.status(200).json(variantType);
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const createVariantType: RouterHandler = async (req, res) => {
-    try {
-        const { name, categoryId } = req.body;
-
-        if (!name || !categoryId) {
-            res.status(400).json({
-                message: 'name and categoryId are required',
-            });
-            return;
-        }
-
-        const categoryExists = await prisma.category.findUnique({
-            where: { id: categoryId },
-        });
-
-        if (!categoryExists) {
-            res.status(404).json({
-                error: 'Category not found while creating VariantType',
-            });
-            return;
-        }
-
-        const newVariantType = await prisma.variantType.create({
-            data: {
-                name,
-                categoryId,
-            },
-        });
-
-        res.status(201).json({
-            message: 'VariantType created successfully',
-            data: newVariantType,
-        });
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const updateVariantTypeById: RouterHandler = async (req, res) => {
-    try {
-        const { id, name, categoryId } = req.body;
-
-        const variantType = await prisma.variantType.update({
-            where: { id },
-            data: {
-                name,
-                categoryId,
-            },
-        });
-
-        if (!variantType) {
-            res.status(404).json({ message: 'VariantType not found' });
-            return;
-        }
-
-        res.status(200).json(variantType);
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const deleteVariantTypeById: RouterHandler = async (req, res) => {
-    try {
-        const { id } = req.params;
-
-        const variantType = await prisma.variantType.findUnique({
-            where: { id },
-        });
-
-        if (!variantType) {
-            res.status(404).json({ message: 'VariantType not found' });
-            return;
-        }
-
-        await prisma.variantType.delete({
-            where: { id },
-        });
-
-        res.status(200).json({
-            message: 'Category deleted successfully',
-            data: variantType,
-        });
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
 
 export const getAllProducts: RouterHandler = async (req, res) => {
     try {
@@ -296,106 +114,6 @@ export const deleteProductById: RouterHandler = async (req, res) => {
         res.status(200).json({
             message: 'Product deleted successfully',
             data: product,
-        });
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const getAllAttributes: RouterHandler = async (req, res) => {
-    try {
-        const attributes = await prisma.productAttribute.findMany();
-
-        res.status(200).json(attributes);
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const getAttributeById: RouterHandler = async (req, res) => {
-    try {
-        const { id } = req.params;
-        const attribute = await prisma.productAttribute.findUnique({
-            where: { id },
-        });
-
-        if (!attribute) {
-            res.status(404).json({ message: 'Attribute not found' });
-            return;
-        }
-
-        res.status(200).json(attribute);
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const createAttribute: RouterHandler = async (req, res) => {
-    try {
-        const { productId, typeId, value } = req.body;
-
-        const newAttribute = await prisma.productAttribute.create({
-            data: {
-                productId,
-                typeId,
-                value,
-            },
-        });
-
-        res.status(201).json({
-            massage: 'Attribute created successfully',
-            data: newAttribute,
-        });
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const updateAttributeById: RouterHandler = async (req, res) => {
-    try {
-        const { id } = req.params;
-        const { productId, typeId, value } = req.body;
-
-        const attribute = await prisma.productAttribute.update({
-            where: { id },
-            data: {
-                productId,
-                typeId,
-                value,
-            },
-        });
-
-        if (!attribute) {
-            res.status(404).json({ message: 'Attribute not found' });
-            return;
-        }
-
-        res.status(200).json(attribute);
-    } catch (error) {
-        ErrorHandler(error, res);
-    }
-};
-
-export const deleteAttributeById: RouterHandler = async (req, res) => {
-    try {
-        const { id } = req.params;
-
-        const attribute = await prisma.productAttribute.findUnique({
-            where: { id },
-        });
-
-        if (!attribute) {
-            res.status(404).json({ message: 'Attribute not found' });
-            return;
-        }
-
-        await prisma.productAttribute.delete({
-            where: { id },
-        });
-
-        res.status(200).json({
-            message: 'Attribute deleted successfully',
-            data: attribute,
         });
     } catch (error) {
         ErrorHandler(error, res);
