+++ server/src/routes/tackle.ts
@@ -1,4 +1,4 @@
-import { Request, RequestHandler, Response, Router } from 'express';
+import { Request, Response, Router } from 'express';
 import {
     createTackle,
     deleteTackle,
@@ -7,19 +7,7 @@ import {
     updateTackle,
 } from '../controllers/tackleController';
 
-// Тип из контроллера, с явным указанием типов из express
 type RouterHandler = (req: Request, res: Response) => Promise<Response>;
-
-// Обёртка для совместимости
-const wrapHandler = (handler: RouterHandler): RequestHandler => {
-    return async (req, res, next) => {
-        try {
-            await handler(req, res);
-        } catch (error) {
-            next(error);
-        }
-    };
-};
 
 const router = Router();
